requirements.txt
substack_api>=1.2.0
requests>=2.31
google-genai>=1.0
anthropic>=0.50
beautifulsoup4>=4.12
pydantic>=2.0
python-dotenv>=1.0
tenacity>=8.0
html2text>=2024.2
markdown2>=2.4
